admin/pyproject.toml
[tool.black]
line-length = 88
target-version = ['py38']

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88

[tool.poetry]
name = "web"
version = "0.1.0"
description = ""
authors = ["Torr1co <fabriziotorrico27@gmail.com>"]
readme = "README.md"
packages = [{include = "web", from = "src"}, {include = "core", from = "src"}]

[tool.poetry.dependencies]
python = "^3.8.10"
flask = "^2.3.3"
pylint = "^2.17.5"
pre-commit = "^3.4.0"
psycopg2-binary = "^2.9.7"
flask-sqlalchemy = "^3.1.1"
python-dotenv = "^1.0.0"
flask-migrate = "^4.0.5"
flask-script = "^2.0.6"
flask-simple-crypt = "^0.3.3"
isort = "^5.12.0"
black = "^23.9.1"
jwt = "^1.3.1"
pyjwt = "^2.8.0"
flask-wtf = "^1.1.1"
flask-mail = "^0.9.1"
flask-session = "^0.5.0"
wtforms = {extras = ["email-validator"], version = "^3.1.0"}
marshmallow = "^3.20.1"
email-validator = "^2.0.0.post2"
flask-cors = "^4.0.0"
flask-oauthlib = "^0.9.6"
authlib = "^1.2.1"


[tool.poetry.group.dev.dependencies]
pytest = "^7.4.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
